packages/plugin-pipecat/pyproject.toml
[project]
name = "neocortex-pipecat"
version = "0.1.0"
description = "Pipecat memory service backed by Neocortex"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pipecat-ai>=0.2.0",
  "requests>=2.31.0",
  "loguru>=0.7.2",
  "pydantic>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
]

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

